+++ packages/jdm-editor/src/components/decision-graph/graph/tab-decision-table.tsx
@@ -1,13 +1,13 @@
-import { Variable } from '@gorules/zen-engine-wasm';
 import type { DragDropManager } from 'dnd-core';
 import React, { useMemo } from 'react';
 import { P, match } from 'ts-pattern';
 
 import { getNodeData } from '../../../helpers/node-data';
-import { useNodeType } from '../../../helpers/node-type';
+import { get } from '../../../helpers/utility';
 import { isWasmAvailable } from '../../../helpers/wasm';
 import type { DecisionTableType } from '../../decision-table';
 import { DecisionTable } from '../../decision-table';
+import type { DecisionTablePermission } from '../../decision-table/context/dt-store.context';
 import { useDecisionGraphActions, useDecisionGraphState } from '../context/dg-store.context';
 import type { NodeDecisionTableData } from '../nodes/specifications/decision-table.specification';
 import type { SimulationTrace, SimulationTraceDataTable } from '../simulator/simulation.types';
@@ -19,46 +19,29 @@ export type TabDecisionTableProps = {
 
 export const TabDecisionTable: React.FC<TabDecisionTableProps> = ({ id, manager }) => {
   const graphActions = useDecisionGraphActions();
-  const nodeType = useNodeType(id, { attachGlobals: false });
-  const { nodeName, nodeTrace, inputData, nodeSnapshot } = useDecisionGraphState(({ simulate, decisionGraph }) => ({
-    nodeName: decisionGraph.nodes.find((n) => n.id === id)?.name,
-    nodeTrace: match(simulate)
-      .with({ result: P.nonNullable }, ({ result }) => result.trace[id] as SimulationTrace<SimulationTraceDataTable>)
-      .otherwise(() => null),
-    inputData: match(simulate)
-      .with({ result: P.nonNullable }, ({ result }) => getNodeData(id, { trace: result.trace, decisionGraph }))
-      .otherwise(() => null),
-    nodeSnapshot: match(simulate)
-      .with(
-        { result: P.nonNullable },
-        ({ result }) => result.snapshot?.nodes?.find((n) => n.id === id)?.content as DecisionTableType,
-      )
-      .otherwise(() => null),
-  }));
-
-  const { disabled, configurable, content, globalType } = useDecisionGraphState(
-    ({ disabled, configurable, decisionGraph, globalType }) => ({
-      disabled,
-      configurable,
-      content: (decisionGraph?.nodes ?? []).find((node) => node.id === id)?.content as NodeDecisionTableData,
-      globalType,
+  const { nodeName, nodeTrace, inputData, nodeSnapshot, viewConfig } = useDecisionGraphState(
+    ({ simulate, decisionGraph, viewConfig }) => ({
+      nodeName: decisionGraph.nodes.find((n) => n.id === id)?.name,
+      nodeTrace: match(simulate)
+        .with({ result: P.nonNullable }, ({ result }) => result.trace[id] as SimulationTrace<SimulationTraceDataTable>)
+        .otherwise(() => null),
+      inputData: match(simulate)
+        .with({ result: P.nonNullable }, ({ result }) => getNodeData(id, { trace: result.trace, decisionGraph }))
+        .otherwise(() => null),
+      nodeSnapshot: match(simulate)
+        .with(
+          { result: P.nonNullable },
+          ({ result }) => result.snapshot.nodes.find((n) => n.id === id)?.content as DecisionTableType,
+        )
+        .otherwise(() => null),
+      viewConfig,
     }),
   );
 
-  const computedType = useMemo(() => {
-    if (!nodeType) {
-      return undefined;
-    }
-
-    const computedType = match(content?.inputField)
-      .with(P.string, (inputField) => nodeType.get(inputField))
-      .otherwise(() => nodeType);
-
-    const newType = content?.executionMode === 'loop' ? computedType.arrayItem() : computedType;
-
-    Object.entries(globalType).forEach(([k, v]) => newType.set(k, v));
-    return newType;
-  }, [nodeType, globalType, content?.inputField, content?.executionMode]);
+  const { disabled, content } = useDecisionGraphState(({ disabled, decisionGraph }) => ({
+    disabled,
+    content: (decisionGraph?.nodes ?? []).find((node) => node.id === id)?.content as NodeDecisionTableData,
+  }));
 
   const debug = useMemo(() => {
     if (!nodeTrace || !inputData || !nodeSnapshot) {
@@ -69,7 +52,12 @@ export const TabDecisionTable: React.FC<TabDecisionTableProps> = ({ id, manager 
       return { trace: nodeTrace, snapshot: nodeSnapshot };
     }
 
-    return { trace: nodeTrace, inputData: new Variable(inputData), snapshot: nodeSnapshot };
+    const extendedInputData = { ...inputData };
+    if (content?.inputField) {
+      extendedInputData.data = get(extendedInputData.data, content.inputField, {});
+    }
+
+    return { trace: nodeTrace, inputData: extendedInputData, snapshot: nodeSnapshot };
   }, [nodeTrace, nodeSnapshot, inputData]);
 
   return (
@@ -80,8 +68,7 @@ export const TabDecisionTable: React.FC<TabDecisionTableProps> = ({ id, manager 
       value={content as any}
       manager={manager}
       disabled={disabled}
-      configurable={configurable}
-      inputData={computedType}
+      permission={viewConfig?.enabled ? (viewConfig?.permissions?.[id] as DecisionTablePermission) : 'edit:full'}
       debug={debug}
       onChange={(val) => {
         graphActions.updateNode(id, (draft) => {