+++ packages/jdm-editor/src/components/decision-graph/graph/tab-expression.tsx
@@ -1,14 +1,15 @@
-import { Variable } from '@gorules/zen-engine-wasm';
 import type { DragDropManager } from 'dnd-core';
 import React, { useMemo } from 'react';
 import { P, match } from 'ts-pattern';
 import type { z } from 'zod';
 
+import type { GetNodeDataResult } from '../../../helpers/node-data';
 import { getNodeData } from '../../../helpers/node-data';
-import { useNodeType } from '../../../helpers/node-type';
 import type { expressionNodeSchema } from '../../../helpers/schema';
+import { get } from '../../../helpers/utility';
 import { isWasmAvailable } from '../../../helpers/wasm';
 import { Expression } from '../../expression';
+import type { ExpressionPermission } from '../../expression/context/expression-store.context';
 import { useDecisionGraphActions, useDecisionGraphState } from '../context/dg-store.context';
 import type { NodeExpressionData } from '../nodes/specifications/expression.specification';
 import type { SimulationTrace, SimulationTraceDataExpression } from '../simulator/simulation.types';
@@ -20,49 +21,32 @@ export type TabExpressionProps = {
 
 export const TabExpression: React.FC<TabExpressionProps> = ({ id, manager }) => {
   const graphActions = useDecisionGraphActions();
-  const nodeType = useNodeType(id, { attachGlobals: false });
-  const { disabled, configurable, content, globalType } = useDecisionGraphState(
-    ({ disabled, configurable, decisionGraph, globalType }) => ({
-      disabled,
-      configurable,
-      globalType,
-      content: (decisionGraph?.nodes ?? []).find((node) => node.id === id)?.content as NodeExpressionData,
-    }),
-  );
-
-  const { nodeTrace, inputData, nodeSnapshot } = useDecisionGraphState(({ simulate, decisionGraph }) => ({
-    nodeTrace: match(simulate)
-      .with(
-        { result: P.nonNullable },
-        ({ result }) => result.trace[id] as SimulationTrace<SimulationTraceDataExpression>,
-      )
-      .otherwise(() => null),
-    inputData: match(simulate)
-      .with({ result: P.nonNullable }, ({ result }) => getNodeData(id, { trace: result.trace, decisionGraph }))
-      .otherwise(() => null),
-    nodeSnapshot: match(simulate)
-      .with(
-        { result: P.nonNullable },
-        ({ result }) =>
-          result.snapshot?.nodes?.find((n) => n.id === id)?.content as z.infer<typeof expressionNodeSchema>['content'],
-      )
-      .otherwise(() => null),
+  const { disabled, content } = useDecisionGraphState(({ disabled, decisionGraph }) => ({
+    disabled,
+    content: (decisionGraph?.nodes ?? []).find((node) => node.id === id)?.content as NodeExpressionData,
   }));
 
-  const computedType = useMemo(() => {
-    if (!nodeType) {
-      return undefined;
-    }
-
-    const computedType = match(content?.inputField)
-      .with(P.string, (inputField) => nodeType.get(inputField))
-      .otherwise(() => nodeType);
-
-    const newType = content?.executionMode === 'loop' ? computedType.arrayItem() : computedType;
-
-    Object.entries(globalType).forEach(([k, v]) => newType.set(k, v));
-    return newType;
-  }, [nodeType, globalType, content?.inputField, content?.executionMode]);
+  const { nodeTrace, inputData, nodeSnapshot, viewConfig } = useDecisionGraphState(
+    ({ simulate, decisionGraph, viewConfig }) => ({
+      nodeTrace: match(simulate)
+        .with(
+          { result: P.nonNullable },
+          ({ result }) => result.trace[id] as SimulationTrace<SimulationTraceDataExpression>,
+        )
+        .otherwise(() => null),
+      inputData: match(simulate)
+        .with({ result: P.nonNullable }, ({ result }) => getNodeData(id, { trace: result.trace, decisionGraph }))
+        .otherwise(() => null),
+      nodeSnapshot: match(simulate)
+        .with(
+          { result: P.nonNullable },
+          ({ result }) =>
+            result.snapshot.nodes.find((n) => n.id === id)?.content as z.infer<typeof expressionNodeSchema>['content'],
+        )
+        .otherwise(() => null),
+      viewConfig,
+    }),
+  );
 
   const debug = useMemo(() => {
     if (!nodeTrace || !inputData || !nodeSnapshot) {
@@ -74,8 +58,16 @@ export const TabExpression: React.FC<TabExpressionProps> = ({ id, manager }) => 
     }
 
     const $data = Object.fromEntries(Object.entries(nodeTrace.traceData).map(([k, v]) => [k, safeJson(v.result)]));
+    const extendedInputData: GetNodeDataResult = {
+      ...inputData,
+      $: $data,
+    };
 
-    return { trace: nodeTrace, inputData: new Variable({ ...inputData, $: $data }), snapshot: nodeSnapshot };
+    if (content?.inputField) {
+      extendedInputData.data = get(extendedInputData.data, content.inputField, {});
+    }
+
+    return { trace: nodeTrace, inputData: extendedInputData, snapshot: nodeSnapshot };
   }, [nodeTrace, nodeSnapshot, inputData]);
 
   return (
@@ -83,9 +75,8 @@ export const TabExpression: React.FC<TabExpressionProps> = ({ id, manager }) => 
       <Expression
         value={content?.expressions}
         disabled={disabled}
-        configurable={configurable}
+        permission={(viewConfig?.enabled ? viewConfig?.permissions?.[id] : 'edit:full') as ExpressionPermission}
         manager={manager}
-        inputData={computedType}
         debug={debug}
         onChange={(val) => {
           graphActions.updateNode(id, (draft) => {