+++ packages/jdm-editor/src/components/decision-table/context/dt-store.context.tsx
@@ -6,6 +6,7 @@ import type { StoreApi, UseBoundStore } from 'zustand';
 import { create } from 'zustand';
 
 import type { SchemaSelectProps } from '../../../helpers/components';
+import { type GetNodeDataResult } from '../../../helpers/node-data';
 import type { SimulationTrace, SimulationTraceDataTable } from '../../decision-graph';
 import type { Diff, DiffMetadata } from '../../decision-graph/dg-types';
 import type { TableCellProps } from '../table/table-default-cell';
@@ -121,6 +122,8 @@ export const parseDecisionTable = (decisionTable?: DecisionTableType) => {
   return dt;
 };
 
+export type DecisionTablePermission = 'edit:full' | 'edit:rules' | 'edit:values';
+
 export type DecisionTableStoreType = {
   state: {
     id?: string;
@@ -129,11 +132,12 @@ export type DecisionTableStoreType = {
     cursor: TableCursor | null;
 
     disabled: boolean;
-    configurable: boolean;
     disableHitPolicy: boolean;
 
     minColWidth: number;
     colWidth: number;
+
+    permission?: DecisionTablePermission;
 
     inputVariableType?: VariableType;
     derivedVariableTypes: Record<string, VariableType>;
@@ -141,11 +145,13 @@ export type DecisionTableStoreType = {
     inputsSchema?: SchemaSelectProps[];
     outputsSchema?: SchemaSelectProps[];
 
+    debugIndex: number;
+    calculatedInputData?: Variable;
+
     debug?: {
       snapshot: DecisionTableType;
       trace: SimulationTrace<SimulationTraceDataTable>;
-      inputData?: Variable;
-      activeRules: string[];
+      inputData?: GetNodeDataResult;
     };
   };
 
@@ -197,7 +203,6 @@ export const DecisionTableProvider: React.FC<React.PropsWithChildren<DecisionTab
         cursor: null,
 
         disabled: false,
-        configurable: true,
         disableHitPolicy: false,
 
         colWidth: 200,
@@ -208,6 +213,7 @@ export const DecisionTableProvider: React.FC<React.PropsWithChildren<DecisionTab
 
         derivedVariableTypes: {},
         inputVariableType: undefined,
+        debugIndex: 0,
         debug: undefined,
       })),
     [],