+++ packages/jdm-editor/src/components/decision-table/dt-empty.tsx
@@ -1,13 +1,14 @@
-import { type Variable, createVariableType } from '@gorules/zen-engine-wasm';
+import { Variable, VariableType } from '@gorules/zen-engine-wasm';
 import equal from 'fast-deep-equal/es6/react';
 import type React from 'react';
 import { useEffect, useRef } from 'react';
-import { P, match } from 'ts-pattern';
 import { useDebouncedCallback } from 'use-debounce';
 
 import type { SchemaSelectProps } from '../../helpers/components';
+import type { GetNodeDataResult } from '../../helpers/node-data';
 import { isWasmAvailable } from '../../helpers/wasm';
 import type { SimulationTrace, SimulationTraceDataTable } from '../decision-graph';
+import type { DecisionTableStoreType } from './context/dt-store.context';
 import {
   type DecisionTableType,
   parseDecisionTable,
@@ -23,17 +24,19 @@ export type DecisionTableEmptyType = {
   defaultValue?: DecisionTableType;
   value?: DecisionTableType;
   disabled?: boolean;
-  configurable?: boolean;
   disableHitPolicy?: boolean;
   cellRenderer?: (props: TableCellProps) => JSX.Element | null | undefined;
   inputsSchema?: SchemaSelectProps[];
   outputsSchema?: SchemaSelectProps[];
-  inputData?: unknown;
-  debug?: { trace: SimulationTrace<SimulationTraceDataTable>; inputData?: Variable; snapshot: DecisionTableType };
+  permission?: DecisionTableStoreType['state']['permission'];
+  debug?: {
+    trace: SimulationTrace<SimulationTraceDataTable>;
+    inputData?: GetNodeDataResult;
+    snapshot: DecisionTableType;
+  };
   minColWidth?: number;
   colWidth?: number;
   onChange?: (val: DecisionTableType) => void;
-  snapshot?: DecisionTableType;
 };
 
 export const DecisionTableEmpty: React.FC<DecisionTableEmptyType> = ({
@@ -42,17 +45,15 @@ export const DecisionTableEmpty: React.FC<DecisionTableEmptyType> = ({
   defaultValue,
   value,
   disabled = false,
-  configurable = true,
   disableHitPolicy = false,
   inputsSchema,
   outputsSchema,
-  inputData,
   debug,
   colWidth,
   minColWidth,
   cellRenderer,
+  permission = 'edit:full',
   onChange,
-  snapshot,
 }) => {
   const mountedRef = useRef(false);
   const { stateStore, listenerStore } = useDecisionTableRaw();
@@ -70,14 +71,14 @@ export const DecisionTableEmpty: React.FC<DecisionTableEmptyType> = ({
       id,
       name,
       disabled,
-      configurable,
       disableHitPolicy,
       inputsSchema,
       outputsSchema,
       colWidth: colWidth || 200,
       minColWidth: minColWidth || 150,
+      permission,
     });
-  }, [id, name, disabled, configurable, disableHitPolicy, inputsSchema, minColWidth, colWidth, outputsSchema]);
+  }, [id, name, disabled, disableHitPolicy, inputsSchema, minColWidth, colWidth, outputsSchema, permission]);
 
   useEffect(() => {
     listenerStore.setState({
@@ -106,32 +107,60 @@ export const DecisionTableEmpty: React.FC<DecisionTableEmptyType> = ({
   }, []);
 
   useEffect(() => {
-    if (!isWasmAvailable()) {
-      return;
-    }
-
-    stateStore.setState({ inputVariableType: createVariableType(inputData) });
-  }, [inputData]);
-
-  useEffect(() => {
     if (!debug) {
       stateStore.setState({ debug: undefined });
       return;
     }
-
-    const activeRules = match(debug.trace.traceData)
-      .with(P.array(), (t) => t.map((d) => d?.rule?._id))
-      .otherwise((t) => [t?.rule?._id]);
 
     stateStore.setState({
       debug: {
         trace: debug.trace,
         snapshot: debug.snapshot,
         inputData: debug.inputData,
-        activeRules,
       },
     });
-  }, [debug, snapshot]);
+  }, [debug]);
+
+  useEffect(() => {
+    if (!isWasmAvailable()) {
+      return;
+    }
+
+    const isLoop = (store: DecisionTableStoreType['state']) => {
+      return store.debug?.snapshot.executionMode === 'loop';
+    };
+
+    const applyDebug = (state: DecisionTableStoreType['state']) => {
+      const inputData = state.debug?.inputData;
+      if (!inputData) {
+        return;
+      }
+
+      const varInputData = new Variable(inputData.data);
+      if (isLoop(state)) {
+        const newInputData = varInputData.get(state.debugIndex).cloneWith('$nodes', inputData.$nodes);
+        stateStore.setState({
+          calculatedInputData: newInputData,
+          inputVariableType: VariableType.fromVariable(newInputData),
+        });
+      } else {
+        const newInputData = varInputData.cloneWith('$nodes', inputData.$nodes);
+        stateStore.setState({
+          calculatedInputData: newInputData,
+          inputVariableType: VariableType.fromVariable(newInputData),
+        });
+      }
+    };
+
+    applyDebug(stateStore.getState());
+    return stateStore.subscribe((state, prevState) => {
+      if (state.debugIndex === prevState.debugIndex && state.debug === prevState.debug) {
+        return;
+      }
+
+      applyDebug(state);
+    });
+  }, []);
 
   return null;
 };