+++ packages/jdm-editor/src/components/expression/context/expression-store.context.tsx
@@ -6,6 +6,7 @@ import type { z } from 'zod';
 import type { StoreApi, UseBoundStore } from 'zustand';
 import { create } from 'zustand';
 
+import { type GetNodeDataResult } from '../../../helpers/node-data';
 import type { expressionNodeSchema } from '../../../helpers/schema';
 import type { SimulationTrace, SimulationTraceDataExpression } from '../../decision-graph';
 import type { DiffMetadata } from '../../decision-graph/dg-types';
@@ -23,9 +24,12 @@ export type ExpressionEntry = {
   _diff?: DiffMetadata;
 };
 
+export type ExpressionPermission = 'edit:full' | 'edit:values' | 'view';
+
 export type ExpressionStore = {
-  configurable: boolean;
   disabled: boolean;
+
+  permission?: ExpressionPermission;
 
   addRowAbove: (index?: number, data?: Partial<ExpressionEntry>) => void;
   addRowBelow: (index?: number, data?: Partial<ExpressionEntry>) => void;
@@ -39,10 +43,12 @@ export type ExpressionStore = {
 
   inputVariableType?: VariableType;
 
+  debugIndex: number;
+  calculatedInputData?: Variable;
   debug?: {
     snapshot: z.infer<typeof expressionNodeSchema>['content'];
     trace: SimulationTrace<SimulationTraceDataExpression>;
-    inputData?: Variable;
+    inputData?: GetNodeDataResult;
   };
 };
 
@@ -63,8 +69,8 @@ export const ExpressionStoreProvider: React.FC<React.PropsWithChildren<Expressio
   const store = useMemo(
     () =>
       create<ExpressionStore>((set) => ({
-        configurable: true,
         disabled: false,
+        debugIndex: 0,
         addRowAbove: (index = 0) => {
           set(
             produce<ExpressionStore>((draft) => {