+++ packages/jdm-editor/src/components/expression/expression.stories.tsx
@@ -11,15 +11,6 @@ const expressionDefault: ExpressionEntry[] = [
   { id: '3', key: 'customer.purchaseTotals', value: 'sum(map(customer.purchases, #.amount))' },
 ];
 
-const expressionDefaultObject = {
-  customer: {
-    firstName: 'John',
-    lastName: 'Doe',
-    tags: ['premium'],
-    purchases: [{ id: '', amount: 100 }],
-  },
-};
-
 const meta: Meta<typeof Expression> = {
   /* 👇 The title prop is optional.
    * See https://storybook.js.org/docs/react/configure/overview#configure-story-loading
@@ -28,14 +19,16 @@ const meta: Meta<typeof Expression> = {
   title: 'Expression',
   component: Expression,
   args: {
-    configurable: true,
     disabled: false,
     defaultValue: expressionDefault,
-    inputData: expressionDefaultObject,
     onChange: fn(),
+    permission: 'edit:full',
   },
   argTypes: {
-    inputData: { control: 'object' },
+    permission: {
+      control: 'select',
+      options: ['edit:full', 'edit:values', 'view'],
+    },
     manager: { table: { disable: true } },
     value: { table: { disable: true } },
   },